move and separate to interpreter
as base.py and values.py
